refactor: Chat関連のコンポーネントをChatディレクトリに移動した
